Fix parsing of multi-options for pycodestyle
This correctly parses options of the form "E302,W405", which before this
was split into individual characters. This resulted in all errors and/or
being considered when ignoring/selecting.

Since specifying just "E" disables all errors, having a list of
characters meant that it always encountered an individual "E" and thus
disabled all errors.