pylama_eradicate: Remove colon to fix ignore
The text of errors from eradicate began with "E800: ", from which the
error number can not be extracted by Error.[1]  This commit removes the
colon, which matches the style from other linters and fixes the problem.
It also adds E8 to test_ignore_select in test_config to cover this.

[1]: https://github.com/klen/pylama/blob/7.6.6/pylama/errors.py#L93

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>