Fix `RunContext.get_params` for `ignore` and `select` linter params
For `ignore` and `select` linter params `RunContext.get_params`
transforms parameter value which is comma separated string to set of
values by splitting that string. Subsequent calls to `get_params` (in
case of cache miss or if another instance of `RunContext` is used) tries
to perform this transformation again and fails because `set` doesn't
have split method.

This commit adds additional check before converting CSV string into a
set -- if the value is already a set, we skip this transformation.